01/31 /Marge/UI_UPdate/toolUpdate/MonsterUpdate
01/31 /Marge/UI_UPdate/toolUpdate/MonsterUpdate

+++ Engine/Utility/Code/Calculator.cpp
@@ -257,7 +257,7 @@ _bool CCalculator::Check_PlayerMoveIndex(const _vec3* pPos, const vector<_int> v
 {
 	_ulong	dwIndex = _ulong(pPos->z / dwVtxItv) * dwCntX + _ulong(pPos->x / dwVtxItv);
 
-	for (int i = 0; i < veciIndex.size(); ++i)
+	for (int i = 0; i < (int)veciIndex.size(); ++i)
 	{
 		if (dwIndex == veciIndex[i])
 			return false;
